fix UB of shift
* the following UBs are reported by ubsan
  + shift exponent 32 is too large for 32-bit type
  + left shift of negative value

+++ include/cpu/decode.h
@@ -14,8 +14,8 @@ typedef struct Decode {
 // --- pattern matching mechanism ---
 __attribute__((always_inline))
 static inline void pattern_decode(const char *str, int len,
-    uint32_t *key, uint32_t *mask, uint32_t *shift) {
-  uint32_t __key = 0, __mask = 0, __shift = 0;
+    uint64_t *key, uint64_t *mask, uint64_t *shift) {
+  uint64_t __key = 0, __mask = 0, __shift = 0;
 #define macro(i) \
   if ((i) >= len) goto finish; \
   else { \
@@ -46,8 +46,8 @@ finish:
 
 __attribute__((always_inline))
 static inline void pattern_decode_hex(const char *str, int len,
-    uint32_t *key, uint32_t *mask, uint32_t *shift) {
-  uint32_t __key = 0, __mask = 0, __shift = 0;
+    uint64_t *key, uint64_t *mask, uint64_t *shift) {
+  uint64_t __key = 0, __mask = 0, __shift = 0;
 #define macro(i) \
   if ((i) >= len) goto finish; \
   else { \
@@ -73,7 +73,7 @@ finish:
 
 // --- pattern matching wrappers for decode ---
 #define INSTPAT(pattern, ...) do { \
-  uint32_t key, mask, shift; \
+  uint64_t key, mask, shift; \
   pattern_decode(pattern, STRLEN(pattern), &key, &mask, &shift); \
   if (((INSTPAT_INST(s) >> shift) & mask) == key) { \
     INSTPAT_MATCH(s, ##__VA_ARGS__); \
